--- components/ExperienceAndReviews.tsx
@@ -1,3 +1,0 @@
-export function ExperienceAndReviews(): JSX.Element {
-  return <div>Experience and reviews</div>
-}
