--- components/ProfileStats.tsx
@@ -1,3 +1,0 @@
-export function ProfileStats(): JSX.Element {
-  return <div>Profile Page</div>
-}
